[Update] Refine ChatRoom
- Refine ChatRoom
- Modify Library Version
- Update Readme and Screenshots
