+++ app/src/main/java/levilin/pokemon/dictionary/ui/screen/chat/component/ChatMessageInputBar.kt
@@ -1,0 +1,85 @@
+package levilin.pokemon.dictionary.ui.screen.chat.component
+
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.Card
+import androidx.compose.material.Icon
+import androidx.compose.material.IconButton
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.OutlinedTextField
+import androidx.compose.material.Text
+import androidx.compose.material.TextFieldDefaults
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.Send
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.input.KeyboardCapitalization
+import androidx.compose.ui.unit.dp
+import levilin.pokemon.dictionary.R
+
+@Composable
+fun ChatMessageInputBar(
+    onSendMessage: (String) -> Unit,
+    resetScroll: () -> Unit = {}
+) {
+    var userMessage by rememberSaveable { mutableStateOf("") }
+
+    Card(
+        modifier = Modifier.fillMaxWidth()
+    ) {
+        Row(
+            modifier = Modifier
+                .padding(16.dp)
+                .fillMaxWidth()
+        ) {
+            OutlinedTextField(
+                modifier = Modifier
+                    .align(Alignment.CenterVertically)
+                    .fillMaxWidth()
+                    .weight(0.85f),
+                value = userMessage,
+                label = { Text(stringResource(R.string.chat_label)) },
+                onValueChange = { userMessage = it },
+                keyboardOptions = KeyboardOptions(
+                    capitalization = KeyboardCapitalization.Sentences,
+                ),
+                colors = TextFieldDefaults.outlinedTextFieldColors(
+                    focusedBorderColor = MaterialTheme.colors.onBackground,
+                    focusedLabelColor = MaterialTheme.colors.onSurface,
+                    unfocusedBorderColor = MaterialTheme.colors.onBackground,
+                    unfocusedLabelColor = MaterialTheme.colors.onSurface,
+                    placeholderColor = MaterialTheme.colors.onBackground
+                )
+            )
+            IconButton(
+                onClick = {
+                    if (userMessage.isNotBlank()) {
+                        onSendMessage(userMessage)
+                        userMessage = ""
+                        resetScroll()
+                    }
+                },
+                modifier = Modifier
+                    .padding(start = 16.dp)
+                    .align(Alignment.CenterVertically)
+                    .fillMaxWidth()
+                    .weight(0.15f)
+            ) {
+                Icon(
+                    modifier = Modifier,
+                    imageVector = Icons.AutoMirrored.Filled.Send,
+                    tint = MaterialTheme.colors.onBackground,
+                    contentDescription = "send_icon"
+                )
+            }
+        }
+    }
+}
